Add files via upload
Air quality monitoring refers to the systematic measurement and assessment of the pollutants present in the air. This process involves the use of various instruments and technologies to detect and analyze pollutants such as particulate matter, volatile organic compounds, carbon monoxide, nitrogen oxides, sulfur dioxide, and ozone. The collected data helps authorities and communities understand the levels of pollution and take necessary actions to improve air quality and safeguard public health.